requirements.txt
scikit-learn
pandas==2.2.3
numpy==2.1.2
xgboost
nltk
PyYAML
dvc
dvc[s3]
mlflow==2.17.0
boto3==1.35.36
awscli
lightgbm==4.5.0
optuna
nltk==3.9.1
spacy
Flask==3.0.3
Flask_Cors==5.0.0
joblib==1.4.2
matplotlib==3.9.2
wordcloud==1.9.3
seaborn==0.13.2